+++ app/Filament/Resources/ContentCalendarSlotResource/Pages/EditContentCalendarSlot.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\ContentCalendarSlotResource\Pages;
+
+use App\Filament\Resources\ContentCalendarSlotResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditContentCalendarSlot extends EditRecord
+{
+    protected static string $resource = ContentCalendarSlotResource::class;
+}